fix assets navigation
